in-out button aaded entry/exit

+++ routes/GateRoutes.js
@@ -26,38 +26,41 @@ gateRouter.post("/CheckQR",(request, response, next) =>
 			console.log("error while fetching user in CheckQR: ");
 			console.log(err);
 		}
+		console.log(userId);
 		//console.log(user);
 		if(user)
 		{
 			console.log(user._id);
-			TempGateRecords.find({userId: user._id}, (err, tempRecord)=>
+
+			const now = new Date();
+			let currDate = date.format(now, 'DD-MM-YYYY');
+			let currTime = date.format(now, 'HH:mm:ss');
+
+			if(request.body.out)
 			{
-				console.log(!(tempRecord[0] == undefined));
-				if(err)
+				console.log(request.body.out);
+				TempGateRecords.find({userId: user._id}, (err, tempRecord)=>
 				{
-					console.log("error while fetching tempRecord: ");
-					console.log(err);
-				}
-				if(!(tempRecord[0] == undefined))
-				{
-					let gateRecord = new GateRecords();
-					const now = new Date();
-					let currDate = date.format(now, 'DD-MM-YYYY');
-					let currTime = date.format(now, 'HH:mm:ss');
-
-					gateRecord.userId = tempRecord[0].userId;
-					gateRecord.outDate = tempRecord[0].outDate;
-					gateRecord.outTime = tempRecord[0].outTime;
-
-					gateRecord.inDate = currDate;
-					gateRecord.inTime = currTime;
-					gateRecord.noOfVisitors = tempRecord[0].noOfVisitors;
-
-					gateRecord.save()
-					.then(result=>
+					console.log(tempRecord);
+					console.log(!(tempRecord[0] == undefined));
+					if(err)
 					{
-						console.log("per. recorded");
-						console.log(result);
+						console.log("error while fetching tempRecord: ");
+						console.log(err);
+					}
+					let gateRecord = new GateRecords();
+
+					gateRecord.userId = user._id;
+					gateRecord.outDate = currDate;
+					gateRecord.outTime = currTime;
+
+					if(!(tempRecord[0] == undefined))
+					{
+						
+						gateRecord.inDate = tempRecord[0].inDate;
+						gateRecord.inTime = tempRecord[0].inTime;
+
+						gateRecord.noOfVisitors = tempRecord[0].noOfVisitors;
 
 						TempGateRecords.deleteOne({userId: userId})
 						.exec()
@@ -69,7 +72,15 @@ gateRouter.post("/CheckQR",(request, response, next) =>
 						{
 							console.log(err);
 						});
+					}
 
+					gateRecord.save()
+					.then(result=>
+					{
+						console.log("per. record inserted");
+						console.log(result);
+
+						
 						return response.render("GateScanQR",
 						{
 							title: "DA Gate Entry/Exit System",
@@ -80,37 +91,37 @@ gateRouter.post("/CheckQR",(request, response, next) =>
 					{
 						console.log(err);
 					});
-				}
-				else
-				{
-					let newTempRecord = new TempGateRecords();
-					const now = new Date();
-					let currDate = date.format(now, 'DD-MM-YYYY');
-					let currTime = date.format(now, 'HH:mm:ss');
-					newTempRecord.userId = user._id;
-					newTempRecord.outDate = currDate;
-					newTempRecord.outTime = currTime;
-					newTempRecord.noOfVisitors = request.body.noOfVisitors;
+					
+				});
+			}
+			else if(request.body.in)
+			{
+				console.log(request.body.in);
 
-					newTempRecord.save((err) =>
+				let newTempRecord = new TempGateRecords();
+
+				newTempRecord.userId = user._id;
+				newTempRecord.inDate = currDate;
+				newTempRecord.inTime = currTime;
+				newTempRecord.noOfVisitors = request.body.noOfVisitors;
+
+				newTempRecord.save()
+				.then(result=>
+				{
+					console.log("Temp recorded");
+					console.log(result);
+
+					return response.render("GateScanQR",
 					{
-						if(err)
-						{
-							console.log("error while saving TempGateRecords");
-							console.log(err);
-						}
-						else
-						{
-							console.log("temp recorded");
-							return response.render("GateScanQR",
-							{
-								title: "DA Gate Entry/Exit System",
-								messages: "Temp Entry recorded successfully"
-							});
-						}
+						title: "DA Gate Entry/Exit System",
+						messages: "Temp. In Entry recorded successfully"
 					});
-				}
-			});
+				})
+				.catch(err=>
+				{
+					console.log(err);
+				});
+			}
 		}
 		else
 		{
@@ -118,7 +129,7 @@ gateRouter.post("/CheckQR",(request, response, next) =>
 			response.render("GateScanQR",
 			{
 				title: "DA Gate Entry/Exit System",
-				messages: "studentId is not valid"
+				messages: "Submission Failed"
 			});
 		}
 	});
@@ -133,14 +144,6 @@ gateRouter.get("/loadGenerateReport",(request, response)=>
 	});
 });
 
-gateRouter.get("/loadGenerateReport",(request, response)=>
-{
-	response.render("GenerateReportForm",
-	{
-		title: "Generate Report from gate reocrds",
-		messages: null
-	});
-});
 
 gateRouter.post("/generateReport",(request, response)=>
 {
@@ -186,13 +189,26 @@ gateRouter.post("/generateReport",(request, response)=>
 			console.log(err);
 		}
 		console.log(startDate, endDate);
-		response.render("DisplayReport",
+		if(result)
 		{
-			title: "Generated Report From Gate Reocrds",
-			messages: result,
-			startDate: startDate,
-			endDate: endDate
-		});
+			response.render("DisplayReport",
+			{
+				title: "Generated Report From Gate Reocrds",
+				messages: result,
+				startDate: startDate,
+				endDate: endDate
+			});
+		}
+		else
+		{
+			response.render("DisplayReport",
+			{
+				title: "Generated Report From Gate Reocrds",
+				messages: "No Data Found",
+				startDate: startDate,
+				endDate: endDate
+			});
+		}
 	});
 });
 
@@ -222,5 +238,93 @@ module.exports = gateRouter;
 		"users.password":0
 	}
 }
+
+*/
+
+/*
+
+code for without in/out button entry/exit
+
+TempGateRecords.find({userId: user._id}, (err, tempRecord)=>
+{
+	console.log(!(tempRecord[0] == undefined));
+	if(err)
+	{
+		console.log("error while fetching tempRecord: ");
+		console.log(err);
+	}
+	if(!(tempRecord[0] == undefined))
+	{
+		let gateRecord = new GateRecords();
+		const now = new Date();
+		let currDate = date.format(now, 'DD-MM-YYYY');
+		let currTime = date.format(now, 'HH:mm:ss');
+
+		gateRecord.userId = tempRecord[0].userId;
+		gateRecord.outDate = tempRecord[0].outDate;
+		gateRecord.outTime = tempRecord[0].outTime;
+
+		gateRecord.inDate = currDate;
+		gateRecord.inTime = currTime;
+		gateRecord.noOfVisitors = tempRecord[0].noOfVisitors;
+
+		gateRecord.save()
+		.then(result=>
+		{
+			console.log("per. recorded");
+			console.log(result);
+
+			TempGateRecords.deleteOne({userId: userId})
+			.exec()
+			.then((result)=>
+			{
+				console.log("temp record delete for "+ userId);
+			})
+			.catch((err)=>
+			{
+				console.log(err);
+			});
+
+			return response.render("GateScanQR",
+			{
+				title: "DA Gate Entry/Exit System",
+				messages: "Per. Entry recorded successfully"
+			});
+		})
+		.catch(err=>
+		{
+			console.log(err);
+		});
+	}
+	else
+	{
+		let newTempRecord = new TempGateRecords();
+		const now = new Date();
+		let currDate = date.format(now, 'DD-MM-YYYY');
+		let currTime = date.format(now, 'HH:mm:ss');
+		newTempRecord.userId = user._id;
+		newTempRecord.outDate = currDate;
+		newTempRecord.outTime = currTime;
+		newTempRecord.noOfVisitors = request.body.noOfVisitors;
+
+		newTempRecord.save((err) =>
+		{
+			if(err)
+			{
+				console.log("error while saving TempGateRecords");
+				console.log(err);
+			}
+			else
+			{
+				console.log("temp recorded");
+				return response.render("GateScanQR",
+				{
+					title: "DA Gate Entry/Exit System",
+					messages: "Temp Entry recorded successfully"
+				});
+			}
+		});
+	}
+});
 
 */
